move common stuff to baseelement
